Object-Oriented Computational Thinking Project
During the COVID-19 pandemic, I witnessed the rapid surge of e-commerce and contactless payments in Latin America and the Caribbean. This significant shift in consumer behavior spurred me into action. Leveraging object-oriented computational thinking, I designed and developed creative solutions to address the information management needs in e-commerce transactions. These solutions not only proved functional but also provided a robust foundation for small and medium-sized enterprises to venture into the online commerce realm. This project not only represented an effective response to the challenges posed by the pandemic but also showcased the power of technological innovation in driving economic growth in the region.